move some files to components/home
