+++ components/home/Column.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import RView, { MediaRule, mergeRStyle } from "emotion-native-media-query";
-import { BREAKPOINTS } from "../helpers/constants";
+import { BREAKPOINTS } from "../../helpers/constants";
 
 export const Column: React.ElementType = (props: any) => {
   const { style = {}, rStyle = {}, ...remainingProps } = props;
