+++ components/home/DesktopRow.tsx
