+++ components/home/ListStyleArticle.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Article, ArticleHeader } from "../Article";
 import { AuthorView } from "./AuthorView";
-import { ArticleProps } from "../../pages/ArticleProps";
+import { ArticleProps } from "./ArticleProps";
 import { ArticleTitleWithLink } from "./ArticleTitleWithLink";
 
 export const ListStyleArticle: React.ElementType = ({ post }: ArticleProps) => {
