+++ components/home/SideThumbnailArticle.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { View } from "react-native";
 import { Article, ArticleHeader } from "../Article";
 import { AuthorView } from "./AuthorView";
-import { ArticleProps } from "../../pages/ArticleProps";
+import { ArticleProps } from "./ArticleProps";
 import { ThumbnailImageWithLink } from "./ThumbnailImageWithLink";
 import { ArticleTitleWithLink } from "./ArticleTitleWithLink";
