+++ components/home/TopThumbnailArticle.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Article, ArticleHeader } from "../Article";
 import { AuthorView } from "./AuthorView";
-import { ArticleProps } from "../../pages/ArticleProps";
+import { ArticleProps } from "./ArticleProps";
 import { ArticleTitleWithLink } from "./ArticleTitleWithLink";
 import { ThumbnailImageWithLink } from "./ThumbnailImageWithLink";
