+++ components/home/getBorderValue.tsx
