+++ pages/index.tsx.rej
@@ -1,0 +1,39 @@
+diff a/pages/index.tsx b/pages/index.tsx	(rejected hunks)
+@@ -2,5 +2 @@ import React from "react";
+-import {
+-  ScrollView,
+-  RefreshControl,
+-  Platform,
+-} from "react-native";
++import { ScrollView, RefreshControl, Platform } from "react-native";
+@@ -21,6 +17 @@ import {
+-import {
+-  getHomeAsync,
+-  getHomeMoreAsync,
+-  Home,
+-  Post,
+-} from "../helpers/wpapi";
++import { getHomeAsync, getHomeMoreAsync, Home, Post } from "../helpers/wpapi";
+@@ -57 +48,3 @@ export interface ArticleProps {
+-export const getBorderValue = (type: "Top" | "Bottom" | "Left" | "Right"): Style => {
++export const getBorderValue = (
++  type: "Top" | "Bottom" | "Left" | "Right",
++): Style => {
+@@ -61,3 +54,3 @@ export const getBorderValue = (type: "Top" | "Bottom" | "Left" | "Right"): Style
+-      [`border${type}Color`]: COLORS.BORDER_COLOR,
+-      [`border${type}Width`]: 1,
+-    };
++        [`border${type}Color`]: COLORS.BORDER_COLOR,
++        [`border${type}Width`]: 1,
++      };
+@@ -237,5 +230,5 @@ export default class IndexPage extends React.Component<IndexProps, IndexState> {
+-                    <>
+-                      <NewsSection />
+-                      <FeaturedSection />
+-                    </>
+-                  )}
++                  <>
++                    <NewsSection />
++                    <FeaturedSection />
++                  </>
++                )}
